Add Spark 3.5 support (based on PR #2467)
Apply full Spark 3.5 support from elastic/elasticsearch-hadoop PR #2467:
- Add spark35scala212/spark35scala213 build variants (keeps Spark 3.4 compat)
- Replace removed RowEncoder(schema) with ExpressionEncoder API
- Add spark-common-utils and spark-sql-api dependencies for Spark 3.5
- Extend SparkVariantPlugin with per-variant source directory support
- Handle StreamingQueryListener.onQueryIdle via reflection + variant subclasses
- Extract StreamingQueryLifecycleListener from test harness
- Set build.snapshot=false to resolve build-tools from Maven Central

Based on: https://github.com/elastic/elasticsearch-hadoop/pull/2467
