+++ gradle.properties
@@ -4,6 +4,7 @@ systemProp.org.gradle.warning.mode=fail
 ## Settings
 org.gradle.daemon=false
 org.gradle.jvmargs=-Xmx4g
+systemProp.build.snapshot=false
 
 # java homes resolved by environment variables
 org.gradle.java.installations.auto-detect=false
@@ -30,6 +31,7 @@ spark20Version = 2.3.0
 spark22Version = 2.2.3
 spark24Version = 2.4.4
 spark30Version = 3.4.3
+spark35Version = 3.5.6
 
 # same as Spark's
 scala210Version = 2.10.7
