+++ spark/sql-30/build.gradle
@@ -9,8 +9,10 @@ apply plugin: 'spark.variants'
 
 sparkVariants {
     capabilityGroup 'org.elasticsearch.spark.sql.variant'
-    setDefaultVariant "spark30scala213", spark30Version, scala213Version
-    addFeatureVariant "spark30scala212", spark30Version, scala212Version
+    setCoreDefaultVariant "spark30scala213", spark30Version, scala213Version
+    addCoreFeatureVariant "spark30scala212", spark30Version, scala212Version
+    addCoreFeatureVariant "spark35scala213", spark35Version, scala213Version
+    addCoreFeatureVariant "spark35scala212", spark35Version, scala212Version
 
     all { SparkVariantPlugin.SparkVariant variant ->
         String scalaCompileTaskName = project.sourceSets
@@ -57,6 +59,14 @@ sparkVariants {
             add(variant.configuration('api'), "org.apache.spark:spark-core_${variant.scalaMajorVersion}:$variant.sparkVersion") {
                 exclude group: 'javax.servlet'
                 exclude group: 'org.apache.hadoop'
+            }
+            if (variant.sparkVersion >= "3.5.0") {
+                add(variant.configuration('implementation'), "org.apache.spark:spark-common-utils_${variant.scalaMajorVersion}:$variant.sparkVersion") {
+                    exclude group: 'org.apache.hadoop'
+                }
+                add(variant.configuration('implementation'), "org.apache.spark:spark-sql-api_${variant.scalaMajorVersion}:$variant.sparkVersion") {
+                    exclude group: 'org.apache.hadoop'
+                }
             }
 
             add(variant.configuration('implementation'), "org.apache.spark:spark-sql_${variant.scalaMajorVersion}:$variant.sparkVersion") {
@@ -197,6 +207,12 @@ tasks.register('copyPoms', Copy) {
     }
     from(tasks.named('generatePomFileForSpark30scala212Publication')) {
         rename 'pom-default.xml', "elasticsearch-spark-30_2.12-${project.getVersion()}.pom"
+    }
+    from(tasks.named('generatePomFileForSpark35scala212Publication')) {
+        rename 'pom-default.xml', "elasticsearch-spark-35_2.12-${project.getVersion()}.pom"
+    }
+    from(tasks.named('generatePomFileForSpark35scala213Publication')) {
+        rename 'pom-default.xml', "elasticsearch-spark-35_2.13-${project.getVersion()}.pom"
     }
     into(new File(project.buildDir, 'distributions'))
 }
