+++ spark/sql-30/src/itest/scala/org/elasticsearch/spark/sql/streaming/StreamingQueryTestHarness.scala
@@ -19,19 +19,13 @@
 
 package org.elasticsearch.spark.sql.streaming
 
-import java.util.UUID
-import java.util.concurrent.CountDownLatch
-import java.util.concurrent.TimeUnit
-import java.util.concurrent.TimeoutException
-import org.apache.spark.sql.Dataset
-import org.apache.spark.sql.Encoder
-import org.apache.spark.sql.SparkSession
-import org.apache.spark.sql.streaming.StreamingQuery
-import org.apache.spark.sql.streaming.StreamingQueryException
-import org.apache.spark.sql.streaming.StreamingQueryListener
+import org.apache.spark.sql.{Dataset, Encoder, SparkSession}
+import org.apache.spark.sql.streaming.{StreamingQuery, StreamingQueryException, StreamingQueryListener}
 import org.elasticsearch.hadoop.util.TestUtils
 import org.elasticsearch.hadoop.util.unit.TimeValue
-import org.junit.Assert
+
+import java.lang.reflect.Method
+import java.util.concurrent.{TimeUnit, TimeoutException}
 
 /**
  * Manages the creation and graceful teardown of a Spark Streaming Query
@@ -62,119 +56,15 @@ class StreamingQueryTestHarness[S <: java.io.Serializable : Encoder](val sparkSe
   private[this] object Running extends State
   private[this] object Closed extends State
 
-  // Listener to 1) ensure no more than a single stream is running at a time, 2) know when we're done processing inputs
-  // and 3) to capture any Exceptions encountered during the execution of the stream.
-  private [this] class StreamingQueryLifecycleListener extends StreamingQueryListener {
 
-    private var uuid: Option[UUID] = None
-
-    private var inputsRequired = 0L
-    private var inputsSeen = 0L
-
-    private var expectingToThrow: Option[Class[_]] = None
-    private var foundExpectedException: Boolean = false
-    private var encounteredException: Option[String] = None
-
-    private var latch = new CountDownLatch(1) // expects just a single batch
-
-    def incrementExpected(): Unit = inputsRequired = inputsRequired + 1
-
-    def setExpectedException(clazz: Class[_]): Unit = {
-      expectingToThrow match {
-        case Some(cls) => throw new IllegalArgumentException(s"Already expecting exception of type [$cls]!")
-        case None => expectingToThrow = Some(clazz)
-      }
-    }
-
-    // Make sure we only ever watch one query at a time.
-    private def captureQueryID(eventId: UUID): Unit = {
-      uuid match {
-        case Some(id) if eventId != id => throw new IllegalStateException("Multiple queries are not supported")
-        case None => uuid = Some(eventId)
-        case _ => // No problem for now
-      }
-    }
-
-    override def onQueryStarted(event: StreamingQueryListener.QueryStartedEvent): Unit = {
-      captureQueryID(event.id)
-    }
-
-    override def onQueryProgress(event: StreamingQueryListener.QueryProgressEvent): Unit = {
-      captureQueryID(event.progress.id)
-
-      // keep track of the number of input rows seen. When we reach the number of expected inputs,
-      // wait for two 0 values to pass before signaling to close
-
-      val rows = event.progress.numInputRows
-      inputsSeen = inputsSeen + rows
-
-      if (inputsSeen == inputsRequired) {
-        if (rows == 0) {
-          // Don't close after meeting the first input level. Wait to make sure we get
-          // one last pass with no new rows processed before signalling.
-          latch.countDown()
-        }
-      } else if (inputsSeen > inputsRequired) {
-        throw new IllegalStateException("Too many inputs encountered. Expected [" + inputsRequired +
-          "] but found [" + inputsSeen + "]")
-      }
-    }
-
-    override def onQueryTerminated(event: StreamingQueryListener.QueryTerminatedEvent): Unit = {
-      try {
-        captureQueryID(event.id)
-
-        encounteredException = event.exception match {
-          case Some(value) =>
-            // This is a whole trace, get everything after the enclosing SparkException (i.e. the original exception + trace)
-            val messageParts = value.split("\\): ")
-            if (messageParts.size > 1) {
-              val nonSparkMessage = messageParts(1)
-              // Ditch the full trace after the first newline
-              val removedNewLine = nonSparkMessage.substring(0, nonSparkMessage.indexOf("\n"))
-              // Split the class name from the exception message and take the class name
-              Some(removedNewLine.substring(0, removedNewLine.indexOf(":")))
-            } else {
-              // Return the original framework error
-              Some(value.substring(0, value.indexOf(":")))
-            }
-          case None => None
-        }
-
-        val expectedExceptionName = expectingToThrow.map(_.getCanonicalName).getOrElse("None")
-
-        foundExpectedException = encounteredException.exists(_.equals(expectedExceptionName))
-      } finally {
-        // signal no matter what to avoid deadlock
-        latch.countDown()
-      }
-    }
-
-    def waitOnComplete(timeValue: TimeValue): Boolean = latch.await(timeValue.millis, TimeUnit.MILLISECONDS)
-
-    def expectAnotherBatch(): Unit = {
-      latch = new CountDownLatch(1)
-    }
-
-    def assertExpectedExceptions(message: Option[String]): Unit = {
-      expectingToThrow match {
-        case Some(exceptionClass) =>
-          if (!foundExpectedException) {
-            encounteredException match {
-              case Some(s) => Assert.fail(s"Expected ${exceptionClass.getCanonicalName} but got $s")
-              case None => Assert.fail(message.getOrElse(s"Expected ${exceptionClass.getCanonicalName} but no Exceptions were thrown"))
-            }
-          }
-        case None =>
-          encounteredException match {
-            case Some(exception) => Assert.fail(s"Expected no exceptions but got $exception")
-            case None => ()
-          }
-      }
+  @transient private val listener: StreamingQueryLifecycleListener = {
+    val supportsIdleEvent = classOf[StreamingQueryListener].getMethods.exists((method: Method) => method.getName.equals("onQueryIdle"))
+    if (supportsIdleEvent) {
+      Class.forName("org.elasticsearch.spark.sql.streaming.StreamingQueryIdleLifecycleListener").getConstructor().newInstance().asInstanceOf[StreamingQueryLifecycleListener]
+    } else {
+      new StreamingQueryLifecycleListener()
     }
   }
-
-  @transient private val listener = new StreamingQueryLifecycleListener
 
   // Port number to use for the socket source
   @transient private lazy val port = {
