+++ spark/sql-30/src/itest/spark35scala213/org/elasticsearch/spark/sql/streaming/StreamingQueryIdleLifecycleListener.scala
@@ -1,0 +1,31 @@
+/*
+ * Licensed to Elasticsearch under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.elasticsearch.spark.sql.streaming
+
+import org.apache.spark.sql.streaming.StreamingQueryListener
+
+// Listener to 1) ensure no more than a single stream is running at a time, 2) know when we're done processing inputs
+// and 3) to capture any Exceptions encountered during the execution of the stream.
+private[streaming] class StreamingQueryIdleLifecycleListener extends StreamingQueryLifecycleListener {
+
+  override def onQueryIdle(event: StreamingQueryListener.QueryIdleEvent): Unit = {
+    super.onQueryIdle(event.id)
+  }
+}
